apps/backend/requirements.txt
altgraph==0.17.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
backports.asyncio.runner==1.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
ConfigArgParse==1.7.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.120.4
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
fsspec==2025.10.0
future==0.18.2
fuzzywuzzy==0.18.0
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
Levenshtein==0.27.1
locust==2.34.0
macholib==1.15.2
MarkupSafe==3.0.3
Metaphone==0.6
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.2.1
numpy==2.0.2
openai==2.7.1
packaging==25.0
parallel-web==0.3.3
pillow==11.3.0
pluggy==1.6.0
postgrest==2.23.0
propcache==0.4.1
psutil==7.1.3
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.2.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.13.0
realtime==2.23.0
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.15.0
sniffio==1.3.1
starlette==0.49.3
storage3==2.23.0
StrEnum==0.4.15
supabase==2.23.0
supabase-auth==2.23.0
supabase-functions==2.23.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
upstash-redis==1.5.0
upstash-search==0.1.1
upstash-vector==0.8.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
